refactor: config 변경 및 디렉토리 변경 (#148)
