+++ src/libs/formatScoreSummary.ts
